+++ sop.h
@@ -18,12 +18,12 @@ void simulation_ctrl();
 void underdamped_ctrl();
 void overdamped_ctrl();
 // void init_native();
-void underdamped_iteration(double3*);
-void overdamped_iteration(double3*);
-void calculate_observables(double3*);
+void underdamped_iteration(float3*);
+void overdamped_iteration(float3*);
+void calculate_observables(float3*);
 // void init_crowder_config();
 // void save_init_crowder_config();
-// void generator_warmup(double);
+// void generator_warmup(float);
 void check_distances();
 
 void run_pair_list_update();
@@ -32,10 +32,10 @@ void run_neighbor_list_update();
 
 void underdamped_update_pos();
 void underdamped_update_pos_gpu();
-__global__ void underdamped_update_pos_kernel(double3 *dev_vel, double3 *dev_force, double3 *dev_pos, double3 *dev_unc_pos, double3 *dev_incr, double a1, double a2, double boxl, int N);
+__global__ void underdamped_update_pos_kernel(float3 *dev_vel, float3 *dev_force, float3 *dev_pos, float3 *dev_unc_pos, float3 *dev_incr, float a1, float a2, float boxl, int N);
 
 void underdamped_update_vel();
 void underdamped_update_vel_gpu();
-__global__ void underdamped_update_vel_kernel(double3 *dev_vel, double3 *dev_force, double3 *dev_incr, double a3, double a4, int N);
+__global__ void underdamped_update_vel_kernel(float3 *dev_vel, float3 *dev_force, float3 *dev_incr, float a3, float a4, int N);
 
 #endif /* SOP_H */
